# codebase for reporting/openapi.yaml
openapi: 3.0.3
info:
  title: NL→SQL Multi-Agent Analytics API
  version: 1.0.0
  description: |
    Backend API for the NL→SQL multi-agent pipeline.

    This service exposes:
    - Agent-0 intent detection
    - Full NL→SQL pipeline (Agents 0–3 + optional DB execution)
    - Direct Oracle SQL execution
    - Report generation (HTML / CSV / PDF / DOCX)

    Frontend (Streamlit) communicates with this API over HTTP only.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /v1/sql/execute:
    post:
      tags: [sql]
      summary: Execute Oracle SQL
      description: |
        Executes a SQL query on the Oracle database using the existing `run_sql` helper.

        Intended primarily for SELECT-style queries; basic guards are in place against
        obviously destructive operations (DROP / DELETE / UPDATE / INSERT / TRUNCATE).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SqlExecuteRequest'
      responses:
        "200":
          description: SQL executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SqlExecuteResponse'
        "400":
          description: Potentially destructive SQL detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Oracle execution error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          description: Error detail message
          oneOf:
            - type: string
            - type: object

    # ---------- Intent ----------
...
    SqlExecuteRequest:
      type: object
      required:
        - sql
      properties:
        sql:
          type: string
          description: Oracle SQL query to execute (SELECT-style only).
          example: SELECT 1 AS test_col FROM DUAL
        row_limit:
          type: integer
          minimum: 1
          maximum: 5000
          default: 200
          description: Maximum number of rows to fetch.
    SqlExecuteResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
          example: ["TEST_COL"]
        rows:
          type: array
          items:
            type: array
            items:
              # values can be of various DB types; represent as any
              nullable: true
          example:
            - [1]
        row_limit:
          type: integer
          example: 200
        exec_ms:
          type: integer
          description: Execution time in milliseconds.
          example: 12

    # ---------- Pipeline ----------
...